scripts/requirements.txt
# Dependencies for News Pipeline Worker (GitHub Actions)
# Install with: pip install -r scripts/requirements.txt

# Firebase Admin SDK (Firestore)
firebase-admin>=6.0.0

# Cloudinary (image upload)
cloudinary>=1.40.0

# HTTP Requests
requests>=2.31.0

# Cerebras Cloud SDK (LLM — GPT-OSS 120B, llama3.1-8b)
cerebras-cloud-sdk>=1.0.0




# Google Gemini Official API (image generation via API key)
google-genai>=1.0.0
Pillow>=10.0.0

# Environment Variables (local dev only)
python-dotenv>=1.0.0
